Fix simple recursion
There are still variable scope issues that need to be fixed, but it passes one
more unit test than it did before. Not touching anything before I commit so I
don't break anything. Still need to sand the bikeshed door so it's a little
less glossy and handle overflows and missing variables.